Extract and unify customization section
