+++ projects/packages/publicize/_inc/components/customize-and-preview/customization-section/styles.module.scss
@@ -1,0 +1,6 @@
+.customization-section {
+	padding: 1rem;
+	display: flex;
+	flex-direction: column;
+	gap: 1.5rem;
+}
