+++ projects/packages/publicize/_inc/components/customize-and-preview/tab-panels/mixins.scss
@@ -22,10 +22,3 @@
 		}
 	}
 }
-
-@mixin customization-section() {
-	padding: 1rem;
-	display: flex;
-	flex-direction: column;
-	gap: 1.5rem;
-}
