+++ projects/packages/publicize/_inc/components/customize-and-preview/tab-panels/mobile/styles.module.scss
@@ -21,11 +21,6 @@
 	grid-template-columns: 1fr;
 }
 
-.customization-section {
-
-	@include mixins.customization-section();
-}
-
 .customization-toggle-wrapper {
 	padding: 1rem;
